wifi config added

+++ usart.c
@@ -3,7 +3,7 @@
 
 #include "usart.h"
 
-void USART1_Init(uint16_t Baudrate , uint16_t Interrupt_Flag , FunctionalState NewState)
+void USART1_Init(uint32_t Baudrate , uint16_t Interrupt_Flag , FunctionalState NewState)
 {
 	// Initialization struct
 	USART_InitTypeDef USART_InitStruct;
