+++ usart.h
@@ -8,7 +8,7 @@
 #include "stm32f10x.h"
 
 
-void USART1_Init(uint16_t Baudrate , uint16_t Interrupt_Flag , FunctionalState NewState);
+void USART1_Init(uint32_t Baudrate , uint16_t Interrupt_Flag , FunctionalState NewState);
 void USART1_PutChar(char c);
 void USART1_PutString(char *s);
 uint16_t USART1_GetChar(void);
